refactor: export TS bindings from a binary instead of a test
The export logic already lives in the plain `export_ts_bindings`
function; only its driver was a test. Run it from a binary gated on
`required-features = ["ts_rs"]` instead, so the export skips the
crate's dev-dependency build and cannot be split across processes by
a test runner. The staging that keeps a failed export from leaving
`TS_RS_EXPORT_DIR` half-written moves into the binary unchanged.
