+++ CLAUDE.md
@@ -109,24 +109,23 @@ need a TS export even if they resemble an exported type.
   see `?` on fields the conductor always serializes, an accepted trade-off —
   do not remove the annotation to "fix" the response side. Response-only
   types keep required fields.
-- Each crate forwards `ts_rs` to the in-set deps it needs;
-  never enable `ts_rs` in a crate's own dev-dependency self-reference, or
-  export tests run during normal `make test-workspace`.
+- Each crate forwards `ts_rs` to the in-set deps it needs; never enable
+  `ts_rs` in a crate's own dev-dependency self-reference, which would drag
+  ts-rs into every `make test-workspace` build.
 - Manual/alias impls are exported by each crate's
   `ts_rs`-gated `export_ts_bindings` function (crate root or `ts` module),
   which chains its in-set deps' functions; when adding one, register it
-  there. The whole tree is written by the single aggregate test in
-  `holochain_conductor_api` (`make ts-bindings`, which runs `cargo test -p
-  holochain_conductor_api --features ts_rs,unstable-countersigning
-  export_bindings_aggregate`) because ts-rs only merges declarations sharing
-  an output file within one process — never run the export through nextest
-  (process per test) or per-crate loops. The test stages the tree in a temp
-  directory and only replaces `TS_RS_EXPORT_DIR` once the export succeeds, so
-  a failed export never leaves it half-written. `make ts-bindings-test` also
-  runs the ts_rs-gated tests that the `export_bindings_aggregate` name
-  filter excludes (the wire-format smoke test and the tag-injection helper's
-  unit tests) — CI runs this target since ordinary `make test-workspace`
-  never enables the `ts_rs` feature.
+  there. The whole tree is written by `holochain_conductor_api`'s
+  `export-ts-bindings` binary (`make ts-bindings`) because ts-rs only merges
+  declarations sharing an output file within one process — never split the
+  export across per-crate loops. The binary is gated on
+  `required-features = ["ts_rs"]`, so builds that don't enable the feature
+  skip it. It stages the tree in a temp directory and only replaces
+  `TS_RS_EXPORT_DIR` once the export succeeds, so a failed export never
+  leaves it half-written. `make ts-bindings-test` runs the export plus the
+  crate's `ts_rs`-gated tests (the wire-format smoke test and the
+  tag-injection helper's unit tests) — CI runs this target since ordinary
+  `make test-workspace` never enables the `ts_rs` feature.
 - The export build enables `unstable-countersigning` on top of `ts_rs`, so
   the countersigning app API and its session state types both reach the
   bindings. Types and fields gated behind the remaining `unstable-*`
