+++ crates/holochain_conductor_api/Cargo.toml
@@ -8,6 +8,13 @@ documentation = "https://docs.rs/holochain_conductor_api"
 authors = ["Holochain Core Dev Team <devcore@holochain.org>"]
 edition = "2021"
 
+# Writes the TypeScript binding tree. Skipped by builds that don't enable
+# `ts_rs`, including `cargo build --workspace`.
+[[bin]]
+name = "export-ts-bindings"
+path = "src/bin/export-ts-bindings.rs"
+required-features = ["ts_rs"]
+
 # reminder - do not use workspace deps
 [dependencies]
 kitsune2_transport_iroh = { version = "0.5.0", optional = true, default-features = false }
